% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_add_ons.R
\name{add_backbar}
\alias{add_backbar}
\title{Add a backbar to a plot.}
\usage{
add_backbar(gg, measure, bg_color = "Gray 4", width = 0.65)
}
\arguments{
\item{gg}{The ggplot object on which to add backbars}

\item{measure}{The column in the original data frame containing the backband measure.}

\item{bg_color}{The backbars color. Default is \code{"Gray 4"}.}

\item{width}{The backbars width Default is \code{0.65}.}
}
\description{
Add an additional set of bars (\code{geom_bars}) to a plot. 
This adds the bars under most existing layers, and isn't added
to the amount of any current bars.
}

