% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_add_ons.R
\name{add_backtext}
\alias{add_backtext}
\title{Add backtext to a plot.}
\usage{
add_backtext(gg, measure, text_size = 3, vertical_adjust = -0.5)
}
\arguments{
\item{gg}{The ggplot object on which to add backbars}

\item{measure}{The column in the original data frame containing the backband measure.}

\item{text_size}{The backtext size. Default is \code{3}.}

\item{vertical_adjust}{The text vertical adjustment. Default is \code{-0.5}.}
}
\description{
Add an additional set of text (\code{geom_text}) to a plot. Works best
when using with add_backbar.
}
