% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_add_ons.R
\name{add_baseband}
\alias{add_baseband}
\title{Add a baseband to a plot.}
\usage{
add_baseband(gg, direction, min, max, bg_color)
}
\arguments{
\item{gg}{The ggplot object on which to add a baseband}

\item{direction}{The direction on which to add the baseband: \code{x}, or \code{y}}

\item{min}{The location to start the band}

\item{max}{The location to end the band}

\item{bg_color}{The background color for the band area. Default is \code{"Green 4"}}
}
\description{
Add a horizontal or vertical background band (\code{geom_rect}) to a plot. 
This adds the band under most existing layers.
}

