% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_add_ons.R
\name{add_facet}
\alias{add_facet}
\title{Add Red Hat facets to a plot}
\usage{
add_facet(gg, facet_col, ncol = NULL, nrow = NULL, scales = "free")
}
\arguments{
\item{gg}{A ggplot object}

\item{facet_col}{The data frame column by which to group the data}

\item{ncol}{The number of columns to have plots}

\item{nrow}{The number of rows of plots}

\item{scales}{Do you want each plot to have the same measure scale ("fixed") or
each to have their own scale ("free")}
}
\description{
After creating your own ggplot, you may want to separate a plot by groups into
two plots. This is done by faceting.
}
