% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_add_ons.R
\name{add_projection}
\alias{add_projection}
\title{Add a confidence interval to a quik line plot}
\usage{
add_projection(gg, projection, confidence, label_final)
}
\arguments{
\item{gg}{The ggplot to be themed}

\item{projection}{The projection (forecast) column}

\item{confidence}{A logical. Should confidence intervals be shown?}

\item{label_final}{A logical. Should the final maximum be labeled?}
}
\description{
After creating a line plot using \code{quik_lines},
this function will add a confidence interval
}
