% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{change_shade}
\alias{change_shade}
\title{Change the shade of the current Red Hat Color}
\usage{
change_shade(hex_color, n)
}
\arguments{
\item{hex_color}{A hexidecimal color for the background}

\item{n}{A positive or negative number to indicate the direction
of the shade change.}
}
\description{
Most colors have six shades. This function provides the color that
is "n" shades away.
}

