% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utilities.R
\name{factor_yearmon}
\alias{factor_yearmon}
\title{Expand the yearmon into a factor}
\usage{
factor_yearmon(df, dimension, measure, facet_by)
}
\arguments{
\item{df}{The data frame to update the dimension}

\item{dimension}{The column containing the values to compare across (e.g. quarters, types)}

\item{measure}{The data frame column with the numerical values to plot}

\item{facet_by}{The data frame column with the facet data (if necessary).}
}
\description{
This will add in any missing dates with zeroes, then change the yearmon
dimension into a factor so that it can be used in order like all other
factor dimensions.
}

