% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utilities.R
\name{factor_zoo}
\alias{factor_zoo}
\title{Expand yearmon or yearqtr into a factor}
\usage{
factor_zoo(df, dimension, measure, facet_by, type)
}
\arguments{
\item{df}{The data frame to update the dimension}

\item{dimension}{The column containing the values to compare across (e.g. quarters, types)}

\item{measure}{The data frame column with the numerical values to plot}

\item{facet_by}{The data frame column with the facet data (if necessary).}

\item{type}{The zoo type, either \code{"yearmon"} or \code{"yearqtr"}.}
}
\description{
This will add in any missing dates with zeroes, then change the yearmon/yearqtr
dimension into a factor so that it can be used in order like all other
factor dimensions.
}

