% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utilities.R
\name{flip_levels}
\alias{flip_levels}
\title{Flip Levels of a factor}
\usage{
flip_levels(data_frame_column)
}
\arguments{
\item{data_frame_column}{The column in your data frame to reverse the levels}
}
\value{
A column to be added to a dataframe
}
\description{
Sometimes plots require levels to be reversed. This function will flip the order
of factors.
}
