% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plot_colors}
\alias{plot_colors}
\title{Red Hat plot colors.}
\format{A data frame with 4 rows and 1 variable:
 \describe{
      \item{positive}{A positive color (currently Green)}
      \item{negative}{A negative color (currently Gold)}
      \item{grid.color}{A gray to color the grid}
      \item{background.fill}{A gray to color plot text}
      \item{text.dark}{A dark gray to color plot text}
      \item{text.light}{A light color (white) to color plot text}
 }}
\source{
internal
}
\usage{
data(plot_colors)
}
\description{
A list containing a standard set of plot colors from
the Red Hat color palette.
}
\examples{
data(plot_colors)

}
\keyword{datasets}

