% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rh_colors}
\alias{rh_colors}
\title{Red Hat branded colors.}
\format{A data frame with 88 rows and 3 variables:
 \describe{
      \item{name}{color name}
      \item{color}{hexidecimal number}
      \item{type}{primary, qualitative, product, gray, or sequential color}
 }}
\source{
internal
}
\usage{
data(rh_colors)
}
\description{
A data frame containing the official color name and their hex code
for use with \url{ggplot2}. Based on
\url{http://brand.redhat.com/elements/color/}
}
\examples{
data(rh_colors)

}
\keyword{datasets}
