% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{pick_text_color}
\alias{pick_text_color}
\title{Pick text color based on background.}
\usage{
pick_text_color(hex_color)
}
\arguments{
\item{hex_color}{A hexidecimal color for the background}
}
\description{
In order to always see the text, function uses Gray 10
or White based on the "luminance" of the background.
Uses W3C guidelines and calculations found on stackoverflow
here: \url{https://stackoverflow.com/questions/3942878/
how-to-decide-font-color-in-white-or-black-depending-on-background-color}
}
