% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{quik_bars}
\alias{quik_bars}
\title{Plot a quik bar plot}
\usage{
quik_bars(df, dimension, measure, groups, position,
                  text_size, alt_text_size, alt_label, facet_by, background,
                  flip_plot, dim_breaks, palette_type, colors,
                  currency, measure_unit, measure_decimal,
                  v.just, text_cutoff, ...)
}
\arguments{
\item{df}{The data frame containing plot data}

\item{dimension}{Usually the x-axis, values like "quarter" or "account type"}

\item{measure}{The column containing numerical values to be plotted}

\item{groups}{If a stacked or dodged bar, the column containing the groups}

\item{position}{If groups, whether to \code{"stack"} (default) or \code{"dodge"} the groups.}

\item{text_size}{A numeric. The size of the text within the bars (default is 3)}

\item{alt_text_size}{A numeric. The size of the alternative text (default is 3)}

\item{alt_label}{The column containing an alternate label (if any)}

\item{facet_by}{The column to add facets to the plot (if necesary)}

\item{background}{If using \code{facet_by}, should background totals be added?}

\item{flip_plot}{A logical. Should the plot coordinates be flipped?}

\item{dim_breaks}{A vector of specific dimensions values that should be labeled.}

\item{palette_type}{A string. Allowed values are \code{"diverging"},
\code{"sequential"}, and \code{"qualitative"}. Default is \code{"qualitative"}.}

\item{colors}{A string. What color should be used for the bars (e.g.
"gray", "red", "dark red", "blue", "dark blue", "light blue", gold",
"green", "purple", "teal")}

\item{currency}{A string, usually \code{$}}

\item{measure_unit}{A string. Can be \code{\%}, \code{K}, or \code{M}}

\item{measure_decimal}{An integer. The number of decimal places to show. Default is 0.}

\item{v.just}{A numeric. The vertical adjustment for the text.}

\item{text_cutoff}{A number. If text values below a certain number should not be included.}

\item{...}{Additional parameters to pass on it facet_wrap (e.g. scales = free, ncol = 1)}
}
\description{
ggquik has specific standards for plots, and
\code{quik_bars} helps your plot follow these.
}
\examples{
# Create a bar plot from morley data
data(morley)
ggq <- quik_bars(morley, dimension = 'Run', measure = 'Speed', groups = 'Expt')
quik_theme(ggq, axis.text = 'y', axis.title = c('x', 'y'))

}

