% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{quik_clean_axis}
\alias{quik_clean_axis}
\title{Clean the axis in a ggplot theme}
\usage{
quik_clean_axis(gg,  axis.title, axis.text)
}
\arguments{
\item{gg}{The ggplot to be themed}

\item{axis.title}{Which axis title labels are
needed:  \code{NULL, 'x', 'y', or c('x', 'y')}}

\item{axis.text}{Which axis text labels are
needed:  \code{NULL, 'x', 'y', or c('x', 'y')}}
}
\description{
Usually called by quik_theme automatically, this function is used to
either update the font for the axis titles and text, or remove them.
It can do this for both axes, or one at a time.
}
