% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{quik_clean_facet}
\alias{quik_clean_facet}
\title{Clean the facet titles in a ggplot theme}
\usage{
quik_clean_facet(gg)
}
\arguments{
\item{gg}{The ggplot to be themed}
}
\description{
Clean the facet titles in a ggplot theme
}
