% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{quik_clean_grid}
\alias{quik_clean_grid}
\title{Clean the grid in a ggplot theme}
\usage{
quik_clean_grid(gg, grid.lines)
}
\arguments{
\item{gg}{The ggplot to be themed}

\item{grid.lines}{Which grid lines are
needed:  \code{NULL, 'x', 'y', or c('x', 'y')}}
}
\description{
Usually called by quik_theme automatically, this function is used to
clear the gray grid background, and either update the grid lines
or remove them. It can do this for both axes, or just one.
}
