% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{quik_legend}
\alias{quik_legend}
\title{Move the legend to the bottom for a ggplot theme}
\usage{
quik_legend(gg, legend_position)
}
\arguments{
\item{gg}{The ggplot to modify the legend}

\item{legend_position}{The position place the legend. One
of three options: \code{"bottom"}, \code{"right"}, or
\code{"none"}. Default is \code{"none"}.}
}
\description{
Move the legend to the bottom for a ggplot theme
}
