% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{quik_lines}
\alias{quik_lines}
\title{Plot a quik line plot}
\usage{
quik_lines(df, dimension, measure, line_groups, palette_type,
                        line_colors, point_size, dim_breaks,
                        facet_col, area, label_size,
                        currency, measure_unit, measure_decimal,
                        baseline)
}
\arguments{
\item{df}{The data frame containing plot data}

\item{dimension}{Usually the x-axis, values like "quarter" or "account type"}

\item{measure}{The column containing numerical values to be plotted}

\item{line_groups}{The column containing the different groups of lines}

\item{palette_type}{A string. Allowed values are \code{"diverging"},}

\item{line_colors}{The the colors to be used for the line(s)/point(s)}

\item{point_size}{A numeric. The size of the points (default is \code{0}, no points)}

\item{dim_breaks}{A vector of specific dimension values that should be labeled.}

\item{facet_col}{The column containing the group to facet the plots (if desired).}

\item{area}{A logical. Should the plot be drawn as an area plot?}

\item{label_size}{The size of the label text size. Default is 3}

\item{currency}{A string, usually \code{$}}

\item{measure_unit}{A string. Can be \code{\%}, \code{K}, or \code{M}}

\item{measure_decimal}{An integer. The number of decimal places to show.}

\item{baseline}{A vector. First value is the intercept line \code{x} or \code{y},
and the second value is the dimension or measure number that should be the baseline.
For example, \code{c('x', 5)}.}
}
\description{
ggquik has specific standards for plots, and
\code{quik_lines} helps your plot follow these.
}

