% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{quik_theme}
\alias{quik_theme}
\title{Add Red Hat theming to a ggplot}
\usage{
quik_theme(gg, clean_grid, clean_axis, clean_facet,
                      legend, transparent, ...)
}
\arguments{
\item{gg}{The ggplot to be themed}

\item{clean_grid}{A logical. Should the plot grid be cleaned?}

\item{clean_axis}{A logical. Should the plot axes be cleaned?}

\item{clean_facet}{A logical. Should the facet titles be cleaned?}

\item{legend}{A string. One of three options: \code{"bottom"},
\code{"right"}, or \code{"none"}. Default is \code{"bottom"}.}

\item{transparent}{A logical. Should the plot be transparent for saving?}

\item{...}{Parameters to pass on to \code{\link{quik_clean_axis}} such
as \code{axis.title} and \code{axis.text}, or to
\code{\link{quik_clean_grid}} such as \code{grid.lines}.}
}
\description{
After creating your own ggplot, or using \code{\link{quik_bars}} or
\code{\link{quik_lines}}, or \code{\link{quik_bullets}},
\code{quik_theme} will help your plot match CEE Analytics standards. By
default, \code{quik_theme} will clean the grid and axes, clean and move
the legend to the bottom, and make all white elements transparent for
universal use. \code{clean_facet} can be set to TRUE to clean ugly
facet titles.
}

