% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{quik_transparent}
\alias{quik_transparent}
\title{Make a ggplot theme transparent}
\usage{
quik_transparent(gg)
}
\arguments{
\item{gg}{The ggplot to be themed}
}
\description{
Make a ggplot theme transparent
}

