% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utilities.R
\name{quiksave}
\alias{quiksave}
\title{Quik save a quik plot as a png file}
\usage{
quiksave(gg, dir, filename, width, height)
}
\arguments{
\item{gg}{The ggplot to be themed}

\item{dir}{The directory to which the plot will be saved}

\item{filename}{The new filename for the plot}

\item{width}{The width of the png (best between 4 and 6)}

\item{height}{The height of the png (best between 2 and 5)}
}
\description{
Quik save a quik plot as a png file
}

