% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{redhat_hex}
\alias{redhat_hex}
\title{Get the hex of a redhat color in rgb}
\usage{
redhat_hex(r = 0, g = 0, b = 0)
}
\arguments{
\item{r}{The red value. Default is 0.}

\item{g}{The green value. Default is 0.}

\item{b}{The blue value. Default is 0.}
}
\description{
Most colors have six shades. This function provides the color that
is "n" shades away.
}

