% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utilities.R
\name{set_decimal}
\alias{set_decimal}
\title{Set the decimal for the plot text}
\usage{
set_decimal(measure, measure_unit)
}
\arguments{
\item{measure}{The df column to be plotted}

\item{measure_unit}{A string. Can be \code{\%}, \code{K}, or \code{M}}
}
\description{
Set the decimal for the plot text
}
