% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_formatting.R
\name{set_group_colors}
\alias{set_group_colors}
\title{Determine quik ggplot2 fill colors}
\usage{
set_group_colors(group_column, palette_type, color_names = NULL)
}
\arguments{
\item{group_column}{The data frame column containing the levels}

\item{palette_type}{A string. Allowed values are \code{"diverging"},
\code{"sequential"}, and \code{"qualitative"}.}

\item{color_names}{A string. If \code{palette_type} is
\code{"sequential"}, what colors should be used for the bars (i.e.
\code{"gray"}, \code{"red"}, \code{"dark red"}, \code{"blue"},
\code{"dark blue"}, \code{"light blue"}, \code{"gold"},
\code{"green"}, \code{"purple"}, \code{"teal"}).
If \code{palette_type} is \code{"manual"},
what colors should be used (e.g. \code{c("Purple 5", "Blue 2")})}
}
\description{
This is a function containing which determines
the fill colors to be used in the plot based
on the type of colors desired. This is a
sub function for \code{\link{quik_bars}} and
\code{\link{quik_bullets}}.
}

