% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utilities.R
\name{set_quik_family}
\alias{set_quik_family}
\title{Set Overpass as the quik family font}
\usage{
set_quik_family()
}
\description{
This is a function will download the Overpass font from
\url{https://fonts.google.com/} if necessary, then set it as 
the font using showtext.
}

