% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_formatting.R
\name{set_quik_opts}
\alias{set_quik_opts}
\title{Set various quik options for plotting}
\usage{
set_quik_opts(df, dimension, measure, groups, palette_type, colors,
  measure_decimal, measure_unit)
}
\arguments{
\item{df}{The data frame containing plot data}

\item{dimension}{The column containing the values to compare across (e.g. quarters, types)}

\item{measure}{The data frame column with the numerical values to plot}

\item{groups}{The data frame column containing the different groups}

\item{palette_type}{A string. Allowed values are \code{"diverging"},
\code{"sequential"}, and \code{"qualitative"}. Default is \code{"qualitative"}.}

\item{colors}{The the colors to be used for the line(s)/point(s)}

\item{measure_decimal}{An integer. The number of decimal places to show.}

\item{measure_unit}{A string. Can be \code{\%}, \code{K}, or \code{M}}
}
\description{
Set various quik options for plotting
}
