% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utilities.R
\name{use_limits}
\alias{use_limits}
\title{Replace ggplot scale breaks with only limits}
\usage{
use_limits()
}
\description{
This is a function containing a function that will take
the given breaks for a continuous scale and returns only
the limits based on the data. Currently only used as a
sub function for \code{\link{quik_bullets}}
}
