% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{quik_bullets}
\alias{quik_bullets}
\title{Plot a quik bullet plot}
\usage{
quik_bullets(df, group_col, range_low, range_high, bar_fill,
                        dotted_line, solid_line, label_size,
                        palette_type, line_colors,
                        currency, measure_unit, ...)
}
\arguments{
\item{df}{The data frame containing plot data}

\item{group_col}{The column containing the different groups to be faceted}

\item{range_low}{The column with the low end of the bar}

\item{range_high}{The column with the high end of the bar}

\item{bar_fill}{The column with the distance to fill the bar (if needed)}

\item{dotted_line}{The column with the dotted line distance (if needed)}

\item{solid_line}{The column with the solid line distance (if needed)}

\item{label_size}{The size of the label text size. Default is 3}

\item{palette_type}{A string. Allowed values are \code{"diverging"},
\code{"sequential"}, and \code{"qualitative"}. Default is \code{"qualitative"}.}

\item{line_colors}{A string. What color should be used for the lines (e.g.
"gray", "red", "dark red", "blue", "dark blue", "light blue", gold",
"green", "purple", "teal")}

\item{currency}{A string, usually \code{$}}

\item{measure_unit}{A string. Can be \code{\%}, \code{K}, or \code{M}}

\item{...}{Parameters to pass on to facet_wrap, such as \code{nrow} or \code{ncol}}
}
\description{
ggquik has specific standards for plots, and
\code{quik_bullets} helps your plot follow these. A bullet plot
\url{http://en.wikipedia.org/wiki/Bullet_graph} is a bar plot with
additional progress indicators. None are required, but the options
are a partial fill, a dotted line, and a solid line.
}
\examples{
# load a sample data frame
df <- data.frame(group = factor(c('Reliability', 'Accuracy', 'Uptime', 'Efficiency')),
                  dotted = sample(90:100, 4), solid = sample(80:100, 4),
                  fill = sample(80:100, 4), low = c(50, 75, 80, 80), high = rep(100, 4))
# create a plot
ggq <- quik_bullets(df, group_col = 'group', range_low = 'low', range_high = 'high',
                    bar_fill = 'fill', solid_line = 'solid', dotted_line = 'dotted')
quik_theme(ggq, axis.text = 'x')

}

